Move render functions to utils folder
