+++ src/utils/renderToMjml.ts
@@ -1,0 +1,6 @@
+import React from "react";
+import ReactDOMServer from "react-dom/server";
+
+export function renderToMjml(email: React.ReactElement) {
+  return ReactDOMServer.renderToStaticMarkup(email);
+}
